requirements.txt
flask
pillow
ultralytics
opencv-python
numpy
